+++ CHANGELOG.md
@@ -8,3 +8,11 @@
 * Now you have time limit for resolve the maximum number of puzzles. Every time you solve one add 30 seconds to the time limit
 * New top score system.
 * New grafics and other changes
+
+##r3:
+* Update LovePotion to the last commits
+* Update PixelOperator8 font 
+* Minor changes on the code (need for run on the last LovePotion)
+* Added more puzzles and credits
+* Update README.md
+* Other minor changes